+++ pkg/server/spec/runner.go
@@ -20,6 +20,7 @@ type RunnerPodBuilder struct {
 	fs         storage.Storage
 	cl         *ContainerLimits
 	labels     Labels
+	csp        *CloudSQLProxy
 }
 
 func (b *RunnerPodBuilder) newAppRunnerContainer() *Container {
@@ -72,6 +73,11 @@ func (b *RunnerPodBuilder) newAppRunnerPod(appContainer *Container) *Pod {
 		builder = builder.WithSideCar(nc, nVol, nCm, SwitchPortWithAppContainer)
 	}
 
+	if b.csp != nil {
+		cn := NewCloudSQLProxyContainer(b.csp)
+		builder = builder.WithSideCar(cn)
+	}
+
 	return builder.Build()
 }
 
@@ -117,6 +123,11 @@ func (b *RunnerPodBuilder) WithLabels(lb Labels) *RunnerPodBuilder {
 
 func (b *RunnerPodBuilder) Build() *Pod {
 	return b.newAppRunnerPod(b.newAppRunnerContainer())
+}
+
+func (b *RunnerPodBuilder) WithCloudSQLProxySideCar(csp *CloudSQLProxy) *RunnerPodBuilder {
+	b.csp = csp
+	return b
 }
 
 func NewRunnerPodBuilder(name, image, initImage string) *RunnerPodBuilder {